requirements.txt
pandas==2.2.0
pillow==10.2.0
tqdm==4.66.1
numpy==1.26.3
requests==2.31.0